플레이리스트 생성, 수정 service 구현

+++ src/main/java/com/musi/shop/web/repository/song/SongRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface SongRepository extends JpaRepository<Song, Integer> {
+public interface SongRepository extends JpaRepository<Song, Long> {
     List<Song> findByAlbumId(Long albumId);
     List<Song> findBySongnameContaining(String keyword); //song 검색
 }
